+++ Samples~/NetworkTemplate/Scripts/Enums/ConnectStatus.cs
@@ -1,0 +1,10 @@
+namespace DSC.Template
+{
+    public enum ConnectStatus
+    {
+        Success,
+        ServerFull,
+        MatchStarted,
+        Unknown,
+    }
+}