+++ Samples~/NetworkTemplate/Scripts/Names/NetMessageName.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DSC.Template
+{
+    public struct NetMessageName
+    {
+        public const string s2c_ConnectResult = "S2C_ConnectResult";
+    }
+}