+++ Scripts/Enums/DSC_NetworkEventType.cs
@@ -2,9 +2,9 @@ namespace DSC.Network
 {
     public enum DSC_NetworkEventType
     {
-        PreStartNetwork,
-        PostStartNetwork,
-        PreStopNetwork,
-        PostStopNetwork
+        StartNetwork,
+        StopNetwork,
+        NetworkStart,
+        TryConnectTimeout,
     }
 }