+++ Scripts/Enums/NetworkMode.cs
@@ -2,8 +2,8 @@ namespace DSC.Network
 {
     public enum NetworkMode
     {
-        Host,
-        Client,
-        Server
+        Host    =   1 << 0,
+        Client  =   1 << 1,
+        Server  =   1 << 2
     }
 }