+++ Scripts/Events/DSC_Event_Network.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace DSC.Network.Event.Helper
+namespace DSC.Network.Event
 {
     public class DSC_Event_Network : MonoBehaviour
     {